Rushed for the deadline achieved nothing...
A draft of a simple PoC for the Graphical Assembler project.
However, this needs to be rewritten with either others libraries,
or a more dynamic language, as C drags everything down when trying
to rush something quickly do demonstrate a concept.

Also, since this whole software must rewrite itself, the first
language is barely relevant anyway...

I guess it served as a good warning : Don't rush !

Signed-off-by: Myy <myy@miouyouyou.fr>